+++ R/mirtCAT_preamble.R
@@ -40,7 +40,8 @@ mirtCAT_preamble <- function(..., final_fun = NULL){
 mirtCAT_preamble_internal <- 
     function(df = NULL, mo = NULL, method = 'MAP', criteria = 'seq', AnswerFuns = list(),
              start_item = 1, local_pattern = NULL, design_elements=FALSE, cl=NULL,
-             design = list(), shinyGUI = list(), preCAT = list(), final_fun = NULL, ...)
+             design = list(), shinyGUI = list(), preCAT = list(), customTypes = list(),
+             final_fun = NULL, ...)
     {
         is_adaptive <- !is.null(mo)
         Names <- if(!is.null(mo)) colnames(mo@Data$data) else NULL
@@ -68,20 +69,22 @@ mirtCAT_preamble_internal <-
                 }
             }, local_pattern=local_pattern, item_options=item_options, mins=mo@Data$mins)
         } else {
-            if(!is.data.frame(df) && !is.list(df))
-                stop('df input must be a data.frame or list', call.=FALSE)
+            if(!is.data.frame(df))
+                stop('df input must be a data.frame', call.=FALSE)
             StemExpression <- if(is.null(df$StemExpression)) logical(length(df$Type))
             else as.logical(df$StemExpression)
             stem_expressions <- rep(NA, length(df$Type))
             stem_expressions[StemExpression] <- df$Question[StemExpression]
-            if(is.data.frame(df)){
-                df <- lapply(df, as.character)
-                for(i in 1L:length(df$Type))
-                    if(StemExpression[i]) df$Question[i] <- ''
-                df$Question <- lapply(df$Question, function(x, fun) shiny::withMathJax(fun(x)),
-                                      fun=shinyGUI$stem_default_format)
+            df <- lapply(df, as.character)
+            for(i in 1L:length(df$Type))
+                if(StemExpression[i]) df$Question[i] <- ''
+            df$Rendered_Question <- lapply(df$Question, function(x, fun) shiny::withMathJax(fun(x)),
+                                  fun=shinyGUI$stem_default_format)
+            if(length(customTypes)){
+                pick <- df$Type %in% names(customTypes)
+                df$Rendered_Question[pick] <- ''
             }
-            obj <- buildShinyElements(df, itemnames = Names)
+            obj <- buildShinyElements(df, itemnames = Names, customTypes=customTypes)
             questions <- obj$questions
             item_answers <- obj$item_answers
             item_options <- obj$item_options
@@ -112,7 +115,7 @@ mirtCAT_preamble_internal <-
         if(!is.null(df)) shinyGUI$stem_locations <- df[['Stem']]
         if(is.null(local_pattern)) 
             shinyGUI_object <- ShinyGUI$new(questions=questions, df=df, shinyGUI=shinyGUI,
-                                            adaptive=is_adaptive)
+                                            adaptive=is_adaptive, CustomTypes=customTypes)
         test_object <- new('Test', mo=mo, item_answers_in=item_answers, AnswerFuns=AnswerFuns,
                            item_options=item_options, quadpts_in=design$quadpts,
                            theta_range_in=design$theta_range, dots=list(...))